fw-lpad/requirements.txt
click==8.0.3
FireWorks==1.9.7
Flask==2.0.2
flask-paginate==2021.12.28
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
monty==2023.11.3
prettytable==3.10.0
pymongo==3.12.3
python-dateutil==2.8.2
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
tabulate==0.8.9
tqdm==4.62.3
wcwidth==0.2.13
Werkzeug==2.0.2
requests==2.26.0
PyYAML==6.0.1

